requirements.txt
Flask==2.3.2
numpy==1.23.5
pandas==1.5.3
opencv-python==4.7.0.72
matplotlib==3.7.1
Pillow==9.5.0
plyfile==0.9
scikit-image==0.20.0
scipy==1.10.1
joblib==1.2.0